# openapi-agent-actions.yaml
...
openapi: 3.1.0
info:
  title: QuranChain Omar Autonomous Agent Actions API
  version: 1.3.0
  description: >
    Granular (single-action) endpoints for an autonomous agent persona ("QuranChain AI" + "Omar AI")
    that can run tasks, create sub-agents, manage integrations, and manage secrets safely.

    Safety model:
    - No password harvesting or "creating credentials" by guessing/inventing logins.
    - External account provisioning is only via approved API/OAuth flows and always requires approval.
    - Secrets can be stored only in a vault; values are never returned.
servers:
  - url: https://api.fullagent.com/v1
    description: Production
  - url: http://localhost:6200/v1
    description: Local Development
...
paths:
  /agents:
    post:
      operationId: createAgent
      tags: [Agent]
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [Agent]
      summary: Get agent
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    patch:
      operationId: updateAgent
      tags: [Agent]
      summary: Update agent (name/description/instructions)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAgentRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{agentId}/personas:
    put:
      operationId: setPersonas
      tags: [Personas]
      summary: Set personas (QuranChain AI, Omar AI)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetPersonasRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{agentId}/specializations:
    put:
      operationId: setSpecializations
      tags: [Specializations]
      summary: Set specializations (QuranChain, DarCloud, Fungi Mesh, Cloudflare, Stripe)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetSpecializationsRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{agentId}/autonomy:
    put:
      operationId: setAutonomy
      tags: [Autonomy]
      summary: Set autonomy policy (approvals, limits)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetAutonomyRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{agentId}/messages:
    post:
      operationId: sendMessage
      tags: [Agent]
      summary: Send message (non-task)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendMessageResponse" }
  /agents/{agentId}/subagents:
    post:
      operationId: createSubAgent
      tags: [SubAgents]
      summary: Create sub-agent
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSubAgentRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubAgent" }

    get:
      operationId: listSubAgents
      tags: [SubAgents]
      summary: List sub-agents
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SubAgent" }
  /agents/{agentId}/subagents/{subAgentId}:
    get:
      operationId: getSubAgent
      tags: [SubAgents]
      summary: Get sub-agent
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/subAgentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubAgent" }

    patch:
      operationId: updateSubAgent
      tags: [SubAgents]
      summary: Update sub-agent
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/subAgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSubAgentRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubAgent" }

    delete:
      operationId: deleteSubAgent
      tags: [SubAgents]
      summary: Delete sub-agent
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/subAgentId"
      responses:
        "204":
          description: Deleted
  /agents/{agentId}/tasks:
    post:
      operationId: createTask
      tags: [Tasks]
      summary: Create autonomous task (goal)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTaskRequest" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }

    get:
      operationId: listTasks
      tags: [Tasks]
      summary: List tasks
      parameters:
        - $ref: "#/components/parameters/agentId"
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/TaskStatus" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
  /agents/{agentId}/tasks/{taskId}:
    get:
      operationId: getTask
      tags: [Tasks]
      summary: Get task
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }

    post:
      operationId: controlTask
      tags: [Tasks]
      summary: Control task (pause/resume/cancel)
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ControlTaskRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /agents/{agentId}/tasks/{taskId}/steps:
    get:
      operationId: listTaskSteps
      tags: [Tasks]
      summary: List task steps
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TaskStep" }
  /agents/{agentId}/tasks/{taskId}/approvals:
    get:
      operationId: getApprovals
      tags: [Approvals]
      summary: Get pending approvals for task
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApprovalRequest" }
  /agents/{agentId}/tasks/{taskId}/approvals/{approvalId}/approve:
    post:
      operationId: approveAction
      tags: [Approvals]
      summary: Approve a pending action
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/taskId"
        - $ref: "#/components/parameters/approvalId"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalDecisionRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /agents/{agentId}/tasks/{taskId}/approvals/{approvalId}/reject:
    post:
      operationId: rejectAction
      tags: [Approvals]
      summary: Reject a pending action
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/taskId"
        - $ref: "#/components/parameters/approvalId"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalDecisionRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
...
  /agents/{agentId}/oauth/providers:
    get:
      operationId: listOauthProviders
      tags: [OAuth]
      summary: List OAuth providers
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OAuthProvider" }
...
components:
...
  schemas:
    CreateAgentRequest:
      type: object
      additionalProperties: false
      required: [name, instructions]
      properties:
        name: { type: string }
        description: { type: string }
        instructions:
          type: string
          description: Include your "QuranChain AI + Omar AI" behavior here if you want it fixed at creation time.
...
    Agent:
      type: object
      additionalProperties: false
      required: [agentId, name]
      properties:
        agentId: { type: string }
        name: { type: string }
        description: { type: string }
        instructions: { type: string }
        personas:
          type: array
          items: { $ref: "#/components/schemas/Persona" }
        specializations:
          type: array
          items: { $ref: "#/components/schemas/Specialization" }
        autonomy:
          $ref: "#/components/schemas/AutonomyPolicy"
...